+++ src/digCsVserver.ts
@@ -1,7 +1,7 @@
 import { deepmergeInto } from "deepmerge-ts";
 import { sortAdcApp } from "./CitrixADC";
 import { logger } from "./logger";
-import { AdcApp, AdcConfObj, AdcRegExTree, PolicyRef } from "./models";
+import { AdcApp, AdcConfObj, AdcRegExTree, Appflow, PolicyRef } from "./models";
 import { parseNsOptions } from "./parseAdc";
 
 
@@ -29,7 +29,7 @@ export async function digCsVservers(coa: AdcConfObj, rx: AdcRegExTree) {
         const opts = parseNsOptions(rxMatch.groups?.opts, rx)
 
         if (!rxMatch) {
-            return logger.error(`regex "${rx.parents["add cs vserver"]}" - failed for line "${originalString}"`);
+            return logger.error(`regex "${rx.parents[parent]}" - failed for line "${originalString}"`);
         }
 
         // object to hold all app details
@@ -108,19 +108,95 @@ export function digAddCsPolicys(app: AdcApp, obj: AdcConfObj, rx: AdcRegExTree) 
                 const originalString = parent + ' ' + x;
                 app.lines.push(originalString);
                 const rxMatch = x.match(rx.parents[parent]);
+                
+                if (!rxMatch) {
+                    return logger.error(`regex "${rx.parents[parent]}" - failed for line "${originalString}"`);
+                }
+                
                 const opts = parseNsOptions(rxMatch.groups.opts, rx)
-
+                
                 // add the policy name to it's details
                 opts.name = rxMatch.groups.name;
-                if(!app.policies) app.policies = [];
-                app.policies.push(opts)
+                if(!app.csPolicies) app.csPolicies = [];
+                app.csPolicies.push(opts)
 
                 if (opts['-action']) {
                     // 'add cs action <name> '
                     // get the action config
                     obj.add.cs.action.filter(el => el.startsWith(opts['-action']))
                         .forEach(x => {
-                            app.lines.push('add cs action ' + x)
+                            const parent = 'add cs action';
+                            const originalString = parent + ' ' + x;
+                            app.lines.push(originalString)
+                            const rxMatch = x.match(rx.parents[parent]);
+                            if (!rxMatch) {
+                                return logger.error(`regex "${rx.parents[parent]}" - failed for line "${originalString}"`);
+                            }
+                            const opts = parseNsOptions(rxMatch.groups.opts, rx)
+                            if(!app.csPolicyActions) app.csPolicyActions = []
+                            app.csPolicyActions.push(opts)
+                        })
+                }
+            })
+
+            //todo:  dig appflow referenced by -policyName
+            // 'add appflow policy <name>' -> 'add appflow action <name>' -> 'add appflow collector <name>'
+
+            obj.add.appflow.policy?.filter(x => x.startsWith(policy['-policyName']))
+            .forEach(x => {
+                //https://developer-docs.citrix.com/projects/netscaler-command-reference/en/12.0/appflow/appflow-policy/appflow-policy/#add-appflow-policy
+                //add appflow policy <name> <rule> <action> [-undefAction <string>] [-comment <string>]
+
+                // applfows are like sflow or Cisco NetFlow.  
+                // not really necessary for deep app parsing to convert to other solutions,
+                //   but having the config lines should provide insight that the feature was configured 
+                //      and may need to be configured in other ways
+
+                const parent = 'add appflow policy';
+                const originalString = parent + ' ' + x;
+                app.lines.push(originalString);
+                const rxMatch = x.match(rx.parents[parent]);
+                const afName = rxMatch.groups.name;
+                const afRule = rxMatch.groups.rule;
+                const afAction = rxMatch.groups.action;
+                
+                if (!rxMatch) {
+                    return logger.error(`regex "${rx.parents[parent]}" - failed for line "${originalString}"`);
+                }
+                
+                // const opts = parseNsOptions(rxMatch.groups.opts, rx)
+
+                const appflow: Appflow = {
+                    name: rxMatch.groups.name,
+                    rule: rxMatch.groups?.rule,
+                    action: []
+                }
+
+                // dig each appflow action                
+
+                if (rxMatch.groups?.action) {
+                    // 'add appflow action <name> '
+                    obj.add.appflow.action.filter(el => el.startsWith(afAction))
+                        .forEach(x => {
+                            const parent = 'add appflow action';
+                            const originalString = parent + ' ' + x;
+                            app.lines.push(originalString)
+                            const rxMatch = x.match(rx.parents[parent]);
+                            if (!rxMatch) {
+                                return logger.error(`regex "${rx.parents[parent]}" - failed for line "${originalString}"`);
+                            }
+                            const opts = parseNsOptions(rxMatch.groups.opts, rx)
+                            if(opts['-collectors']) {
+                                const collectorName = opts['-collectors'];
+                                obj.add.appflow.collector.filter(el => el.startsWith(collectorName))
+                                    .forEach(x => {
+                                        const parent = 'add appflow collector';
+                                        const originalString = parent + ' ' + x;
+                                        app.lines.push(originalString)
+                                    })
+                            }
+                            if(!app.csPolicyActions) app.csPolicyActions = []
+                            // app.csPolicyActions.push(opts)
                         })
                 }
             })